ajuste de conteudo

+++ src/components/listagem/ListagemPaciente.vue
@@ -60,38 +60,55 @@
 
     <q-dialog v-model="formEmEdicao">
       <q-card>
-        <div class="flex">
-          <div class="text-overline text-grey">Nome:</div>
-          <q-input v-model="pacienteEdicao.NOMEPACI" />
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Data de Nascimento:</div>
-          <q-input v-model="pacienteEdicao.DATANASPACI"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Genero:</div>
-          <q-input v-model="pacienteEdicao.GENPACI"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Cidade:</div>
-          <q-input v-model="pacienteEdicao.CIDADEPACI"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">UF:</div>
-          <q-input v-model="pacienteEdicao.UFPACI"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Endereço:</div>
-          <q-input v-model="pacienteEdicao.ENDERPACI"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Telefone:</div>
-          <q-input v-model="pacienteEdicao.TELMED"></q-input>
-        </div>
-        <div class="flex">
-          <div class="text-overline text-grey">Email:</div>
-          <q-input v-model="pacienteEdicao.EMAILMED"></q-input>
-        </div>
+        <q-card-section>
+          <div class="flex">
+            <div class="text-overline text-grey">Nome:</div>
+            <q-input v-model="pacienteEdicao.NOMEPACI" />
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Data de Nascimento:</div>
+            <q-input v-model="pacienteEdicao.DATANASPACI"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Genero:</div>
+            <q-input v-model="pacienteEdicao.GENPACI"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Cidade:</div>
+            <q-input v-model="pacienteEdicao.CIDADEPACI"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">UF:</div>
+            <q-input v-model="pacienteEdicao.UFPACI"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Endereço:</div>
+            <q-input v-model="pacienteEdicao.ENDERPACI"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Telefone:</div>
+            <q-input v-model="pacienteEdicao.TELMED"></q-input>
+          </div>
+          <div class="flex">
+            <div class="text-overline text-grey">Email:</div>
+            <q-input v-model="pacienteEdicao.EMAILMED"></q-input></div
+        ></q-card-section>
+        <q-card-section
+          ><q-btn
+            flat
+            round
+            color="primary"
+            label="Editar"
+            @click="patchPaciente(pacienteEdicao)"
+          />
+          <q-btn
+            flat
+            round
+            color="red"
+            label="Cancelar"
+            @Click="formEmEdicao = false"
+          />
+        </q-card-section>
       </q-card>
     </q-dialog>
   </div>
@@ -100,14 +117,17 @@
 <script>
 import { defineComponent, onMounted, ref } from "vue";
 import api from "../../service/index";
+import { useQuasar } from "quasar";
 
 export default defineComponent({
   name: "ListagemPaciente",
 
   setup() {
+    const q = useQuasar();
     const pacienteEdicao = ref({});
     const pacientes = ref([]);
     const formEmEdicao = ref(false);
+    const desabilitaExclusao = ref(false);
 
     const fetchPacientes = async () =>
       await api.get("/pacientes").then((res) => (pacientes.value = res.data));
@@ -115,7 +135,48 @@ export default defineComponent({
     onMounted(fetchPacientes);
 
     const deletePaciente = async (id) => {
-      await api.delete(`/pacientes/${id}`).then(() => fetchPacientes());
+      await getPacientesComConsultas(id);
+
+      if (desabilitaExclusao.value) {
+        q.notify({
+          color: "negative",
+          message: "Paciente possui consultas agendadas",
+        });
+        return;
+      } else {
+        await api.delete(`/pacientes/${id}`).then(() => fetchPacientes());
+      }
+    };
+
+    const getPacientesComConsultas = async function (id) {
+      try {
+        const response = await api.get(`consultas/verificapaciente/${id}`);
+
+        if (response.data.length > 0) {
+          desabilitaExclusao.value = true;
+          return;
+        }
+      } catch (error) {
+        console.error(error);
+        q.notify({
+          color: "negative",
+          message: "Erro ao verificar paciente",
+        });
+      }
+    };
+
+    const patchPaciente = async function (paciente) {
+      try {
+        const response = await api
+          .patch(`/pacientes/${paciente.IDPACIENTE}`, paciente)
+          .then(() => fetchPacientes());
+      } catch (error) {
+        console.error(error);
+        q.notify({
+          color: "negative",
+          message: "Erro ao editar paciente",
+        });
+      }
     };
 
     const abreEdicao = () => {};
@@ -125,6 +186,7 @@ export default defineComponent({
       pacienteEdicao,
       formEmEdicao,
       abreEdicao,
+      patchPaciente,
     };
   },
 });
